requirements.txt
fastapi==0.115.11
googlemaps==4.10.0
joblib==1.4.2
jsonify==0.5
numpy==2.2.3
pandas==2.2.3
pydantic==2.10.6
pymongo==4.11.2
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.6.1
scikit_surprise==1.1.4
bcrypt==4.0.1